Continue implementation, re-arrange modules.
