+++ erdjs/src/smartcontracts/typesystem/interfaces.ts
@@ -1,7 +1,5 @@
-import { TypeDescriptor } from "./typeDescriptor";
 import { Type } from "./types";
 
 export interface ITypeResolver {
-    resolveTypeDescriptor(scopedTypeNames: string[]): TypeDescriptor;
-    resolveType(typeName: string): Type;
+    resolveType(typeName: string): Type | null;
 }
