+++ erdjs/src/smartcontracts/typesystem/typeResolvers.ts
@@ -1,9 +1,13 @@
+import { AddressType } from "./address";
+import { BooleanType } from "./boolean";
+import { OptionalType, VectorType } from "./generic";
 import { ITypeResolver } from "./interfaces";
 import { TypeDescriptor } from "./typeDescriptor";
 import { Type } from "./types";
 
 export class TypeResolvers implements ITypeResolver {
     private static instance: TypeResolvers;
+    private readonly resolvers: ITypeResolver[] = [];
 
     /**
      * Gets the singleton.
@@ -17,7 +21,11 @@ export class TypeResolvers implements ITypeResolver {
     }
 
     addResolver(resolver: ITypeResolver) {
+        this.resolvers.push(resolver);
+    }
 
+    clearResolvers() {
+        this.resolvers.splice(0, this.resolvers.length);
     }
 
     resolveTypeDescriptor(scopedTypeNames: string[]): TypeDescriptor {
@@ -30,9 +38,20 @@ export class TypeResolvers implements ITypeResolver {
 
         return new TypeDescriptor(types);
     }
-
     
     resolveType(typeName: string): Type {
-        throw new Error("Method not implemented.");
+        // TODO:
+        if (typeName == "Optional") {
+            return new OptionalType();
+        }
+        if (typeName == "Vector") {
+            return new VectorType();
+        }
+        if (typeName == "Boolean") {
+            return new BooleanType();
+        }
+        if (typeName == "Address") {
+            return new AddressType();
+        }
     }
 }
